moved things to userutils class
